requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
httpx>=0.24.0
PyYAML>=6.0
PyJWT[crypto]>=2.8.0
sqlmodel>=0.0.16
bcrypt>=4.0.0
stripe>=7.0.0
